\alias{ConstantExpr}
\name{ConstantExpr}

\alias{isGEPWithNoNotionalOverIndexing}


\title{Constant Expression}
\description{
These are functions for working \code{\link{ConstantExpr-class}} objects.
}
\usage{
isGEPWithNoNotionalOverIndexing(x)
}
\arguments{
  \item{x}{a \code{\link{ConstantExpr-class}} object}
}

\author{DTL}
