\name{getAllUsers}
\alias{getAllUsers}
\alias{getAllUses}
\alias{getUser}


\alias{replaceUsesOfWith}
\alias{replaceAllUsesWith}


\title{Get Uses and Users of a Value}
\description{
}
\usage{
getAllUsers(obj, uses = getAllUses(obj))
}
\arguments{
  \item{obj}{the Value object}
  \item{uses}{a list of ....} %XXX
}
\details{
}
\value{
}
\references{
}
\author{
DTL
}

\seealso{
}
\examples{
}
\keyword{programming}
