\name{getClassName}
\alias{getClassName}
\title{Get the name of a Value's class}
\description{
}
\usage{
getClassName(value)
}
\arguments{
  \item{value}{an LLVM Value object}
}

\value{
 A character string, i.e., a character vector of length 1.
}
\author{
DTL
}
\seealso{
}
\examples{
}
\keyword{programming}
