\name{getMetadata}
\alias{getMetadata}
\alias{getMetadata,Function,character-method}
\alias{getMetadata,Function,numeric-method}
\alias{getMetadata,Instruction,character-method}
\alias{getMetadata,Instruction,numeric-method}
%XX For global variables

\alias{coerce,Metadata,character-method}

\alias{getName,NamedMDNode-method}


\alias{Metadata-class}

\title{Get Meta-data from Functions, Instructions, Global Variables}

\description{
}
\usage{
getMetadata(obj, id, ...)
}
\arguments{
  \item{obj}{}
  \item{id}{}
  \item{\dots}{additional parameters for methods}
}
\value{
}
\references{
}
\author{
DTL
}
\seealso{
}
\examples{
}
\keyword{programming}
