\name{getOpcode}
\alias{getOpcode}
\title{Get the operation code/identifier for an LLVM instruction}
\description{
  This function allows us to retrieve the identifier or type of the
  instruction.
}
\usage{
getOpcode(x, ...)
}
\arguments{
  \item{x}{the instruction}
  \item{\dots}{additional arguments for methods}
}

%\value{}
\references{
  LLVM Documentation \url{http://llvm.org/docs/}
}
\author{Duncan Temple Lang}

%\seealso{}
%\examples{}
\keyword{programming} 
