\name{isInBounds}
\alias{isInBounds}
\alias{isInBounds<-}
\alias{isInBounds,Value-method}
\alias{isInBounds<-,Value-method}

\title{Query or set bounds checking flag}
\description{

}
\usage{
isInBounds(x, ...)
}

\arguments{
  \item{x}{the LLVM \code{Value} object}
  \item{\dots}{additional arguments for methods}
}
\value{
}
%\references{}
\author{

}
\seealso{}
\examples{
}
\keyword{programming}

