\name{isZeroValue}
\alias{isZeroValue}
\alias{isZeroValue.Constant}
\alias{isZeroValue,Constant-method}

\title{Check if a Value is a Zero-initializer}
\description{
}
\usage{
isZeroValue(x, ...)
}
\arguments{
  \item{x}{the LLVM Value object}
  \item{\dots}{arguments for methods}
}

\value{
A logical scalar value.
}
\references{
}
\author{DTL}


\seealso{
}
\examples{
}
\keyword{programming}
