\name{Loop}
\alias{Loop-class}

\alias{getGuardBranch}
\alias{isLoopInvariant}
\alias{hasLoopInvariantOperands}
\alias{isRotatedForm}
\alias{isLoopSimplifyForm}
\alias{isGuarded}
\alias{isAnnotatedParallel}
\alias{getIncomingAndBackEdge}

\title{Methods for \code{Loop} Objects}

\usage{
isAnnotatedParallel(loop)
isRotatedForm(loop)
isLoopSimplifyForm(loop)
isGuarded(loop)
getGuardBranch(loop)
getIncomingAndBackEdge(loop)
isLoopInvariant(loop, val)
hasLoopInvariantOperands(loop, val)
}
