\name{newDebugBasicType}
\alias{newDebugBasicType}
\alias{newDebugCompileUnit}
\alias{newDebugFunction}
\alias{newDebugFunctionType}
\alias{newDebugLocalVariable}
\alias{newDebugPointerType}


\alias{DIBuilder}

\alias{DIBuilder-class}
\alias{DICompileUnit-class}
\alias{DISubroutineType-class}
\alias{DIType-class}
\alias{DISubprogram-class}

\alias{finalizeDIBuilder}%XXX




\alias{isLittleEndian,DIType-method}
\alias{isLValueReference,DIType-method}
\alias{isObjcClassComplete,DIType-method}
\alias{isObjectPointer,DIType-method}
\alias{isPrivate,DIType-method}
\alias{isProtected,DIType-method}
\alias{isPublic,DIType-method}
\alias{isRValueReference,DIType-method}
\alias{isStaticMember,DIType-method}
\alias{isTypePassByReference,DIType-method}
\alias{isTypePassByValue,DIType-method}
\alias{isVector,DIType-method}
\alias{isVirtual,DIType-method}
\alias{isForwardDecl,DIType-method}
\alias{isBigEndian,DIType-method}
\alias{isBitField,DIType-method}
\alias{isAppleBlockExtension,DIType-method}
\alias{isArtificial,DIType-method}
\alias{isAtomic,Instruction-method}



\alias{isDebugDirectivesOnly}
\alias{getProducer}


%- Also NEED an '\alias' for EACH other topic documented here.
\title{Debug Information Functions for DIBuilder} 

\description{

}
\usage{
newDebugBasicType(builder, name, size, align, dwarf_type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{builder}{the DBuilder instance}
  \item{name}{}
  \item{size}{}
  \item{align}{}
  \item{dwarf_type}{}
}
\value{

}
\references{
\url{https://llvm.org/doxygen/classllvm_1_1DIBuilder.html}
}
\author{
DTL
}

\seealso{

}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}

