\name{onlyReadsMemory}
\alias{onlyReadsMemory}
\alias{onlyReadsMemory,Argument-method}
\alias{onlyReadsMemory,Function-method}

\title{Determines if the LLVM does not write to memory}
\description{
}
\usage{
onlyReadsMemory(x, ...)
}
\arguments{
  \item{x}{the LLVM object, currently a \code{Function} or \code{Argument} object}
  \item{\dots}{additional arguments for methods}
}
\value{
}
\references{}
\author{
}

\seealso{}
\examples{
}

\keyword{programming}

