\name{setInitializer}
\alias{setInitializer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Set the initializer for an LLVM Value}
\description{
This function sets the initial value for an LLVM variable or Value object.
}
\usage{
setInitializer(var, value)
}
\arguments{
  \item{var}{the variable whose value is to be set}
  \item{value}{the value with which to initialize the variable}
}
%\value{}
\references{
  LLVM Documentation \url{http://llvm.org/docs/}
}
\author{
DTL
}


%\seealso{}
%\examples{}
\keyword{programming}
