\name{VoidType}
\alias{VoidType}
\alias{DoublePtrType}
\alias{DoubleType}
\alias{FloatPtrType}
\alias{FloatType}
\alias{Int16Type}
\alias{Int1Type}
\alias{Int32PtrType}
\alias{Int32Type}
\alias{Int64Type}
\alias{Int8Type}
\alias{isBasicType}
\alias{LabelType}
\alias{pointerType}

\docType{data}
\title{Objects representing different LLVM common Types}
\description{
  These are a collection of internal objects that represent
  standard data types used by LLVM. These are different
  types of integers and real numbers, pointers, etc.
  We use these to specify types for return values, parameters,
  and different inputs to instructions.
}
\usage{VoidType}
\format{
  The format is:
<externalptr> 
}

\source{
}
\references{
}
\examples{
}
\keyword{datasets}
